// src/app/core/enums/api.enum.ts
export enum Api {
  GetAllCategories = '/Billing/Category/GetAllCategory',
  GetActiveCategories = '/Billing/Category/GetCategory',
  CreateCategory = '/Billing/Category/AddCategory',
  UpdateCategory = '/Billing/Category/UpdateCategory',
  ChangeStatusCategory = '/Billing/Category/ChangeStatusCategory',
  GetAllLocations = '/Billing/Location/GetAllLocation',
  GetActiveLocations = '/Billing/Location/GetLocation',
  CreateLocation = '/Billing/Location/AddLocation',
  UpdateLocation = '/Billing/Location/UpdateLocation',
  ChangeStatusLocation = '/Billing/Location/ChangeStatusLocation',
  GetAllProducts = '/Billing/Item/GetAllItem',
  GetActiveProducts = '/Billing/Item/GetItem',
  CreateProduct = '/Billing/Item/AddItem',
  UpdateProduct = '/Billing/Item/UpdateItem',
  ChangeStatusProduct = '/Billing/Item/ChangeStatusItem',
  GetCustomers = '/Billing/GetCustomers',
  CreateCustomer = '/Billing/AddCustomer',
  UpdateCustomer = '/Billing/UpdateCustomer',
  GetAllProviders = '/Billing/Provider/GetAllProvider',
  GetActiveProviders = '/Billing/Provider/GetProvider',
  CreateProvider = '/Billing/Provider/AddProvider',
  UpdateProvider = '/Billing/Provider/UpdateProvider',
  ChangeStatusProvider = '/Billing/Provider/ChangeStatusProvider',
  GetMigratedSales = '/Billing/GetBillingByCustomerMigration',
  GetAllResolutions = '/Billing/Resolution/GetAllResolution',
  CreateResolution = '/Billing/Resolution/AddResolution',
  UpdateResolution = '/Billing/Resolution/UpdateResolution',
  AddSale = '/Billing/Sales/AddSale',
  PaySale = '/Billing/Sales/PaySale',
  GetSales = '/Billing/Sales/GetSales',
  SearchSales = '/Billing/Sales/GetSalesHistory',
  SearchSalesByCustomer = '/Billing/Sales/GetAllSalesByCustomer',
  AddPurchase = '/Billing/Purchase/AddPurchase',
  SearchPurchases = '/Billing/Purchase/GetPurchaseHistory',
  AddOutput = '/Billing/Output/AddOutput',
  ChangeOutputStatus = '/Billing/Output/ChangeStatusOutput',
  GetOutputs = '/Billing/Output/GetOutputHistory',
  AddBalance = '/Billing/Balance/AddBalance',
  GetBalance = '/Billing/Balance/GetBalanceHistory',
  ChangeBalanceStatus = '/Billing/Balance/ChangeStatusBalance',
  StartBalanceByLocation = '/Billing/Balance/StartBalanceByLocation',
  EndBalanceByLocation = '/Billing/Balance/EndBalanceByLocation',
  ValidateDocument = '/Billing/Document/ValidateDocument',
  DocumentGenerated = '/Billing/Document/DocumentGenerated',
  CancelDocument = '/Billing/Document/CancellDocumentGenerated',
  GetMunicipalities = 'assets/jsons/municipalities.json',
  GetCodeActivities = 'assets/jsons/codeActivities.json',
  SearchPendingPayments = '/Billing/Payment/GetPendingPayments',
  AddPayPendingPayments = '/Billing/Payment/AddPayPendingPayments',
  AddSaleInvoiceDte = '/Billing/Dte/Factura',
  AddSaleTaxCreditDte = '/Billing/Dte/CreditoFiscal',
  CancelDocumentDte = '/Billing/Dte/Invalidacion',
  SendEmail = '/Billing/Dte/SendEmail',
}