// src/app/views/expenses/models/output.model.ts
export interface Output {
  active: boolean;
  amount_output: number;
  comment?: string;
  created_at?: string;
  description: string;
  id: number;
  location?: string;
  location_id?: number;
  output_receiver: string;
  output_type: string;
  update_at?: string;
  user_id?: string;
  user_name?: string;
}